+++ app/src/main/java/com/whc/cvccmeasuresystem/Model/HumidityVO.java
@@ -4,7 +4,7 @@ import java.sql.Timestamp;
 
 public class HumidityVO {
 
-    private Integer id;
+    private Long id;
     private String name;
     private Timestamp beginTime;
     private Integer baseVoltage;
@@ -36,11 +36,12 @@ public class HumidityVO {
         this.overVoltage = overVoltage;
     }
 
-    public Integer getId() {
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
